fix bug cause by moving GenomeUtils.
